family members view refactor
