+++ src/components/atoms/TransparentInput/TransparentInput.js
