+++ src/components/molecules/TasksAddField/TasksAddField.js
@@ -3,7 +3,7 @@ import { useCollectionQueries } from 'hooks/useCollectionQueries'
 import { AddTaskBox, AddBtn } from './TasksAddField.styles'
 import { Timestamp } from 'firebase/firestore'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { TransparentInput } from 'components/atoms/TransparentInput/TransparentInput.styles'
+import { TransparentInput } from 'components/atoms/TransparentInput/TransparentInput'
 import { FamilyContext } from 'providers/CurrentFamilyProvider'
 import { UserContext } from 'providers/ActiveUserProvider'
 import { faPlus } from '@fortawesome/free-solid-svg-icons'
